Move caller_file_and_line to chia._tests.util.misc
